refactor expect events and remove dead code
